requirements.txt
Flask==3.0.3
WTForms==3.0.1
line-bot-sdk==2.0.1
gunicorn==20.0.4
python-dotenv==0.19.0
PyMongo==3.12.1
dnspython==2.1.0
webassets==2.0
pyscss==1.3.7
Authlib==1.6.6
Werkzeug==3.0.1
